Move item search provider registration to separate file.
