+++ lib/Models/IndexedItemSearchProviders.ts
@@ -1,0 +1,15 @@
+import Constructor from "../Core/Constructor";
+import ItemSearchProvider from "./ItemSearchProvider";
+import IndexedItemSearchProvider from "./ItemSearchProviders/IndexedItemSearchProvider";
+
+export const ItemSearchProviders: Map<
+  string,
+  Constructor<ItemSearchProvider>
+> = new Map([["indexed", IndexedItemSearchProvider]]);
+
+export function registerItemSearchProvider(
+  type: string,
+  providerClass: Constructor<ItemSearchProvider>
+) {
+  ItemSearchProviders.set(type, providerClass);
+}
